# Performance tests: Log timings of performance test segments