> **Note**  This PR is not ready to merge but it needs to also not be a draft so that the normal CI steps run. The purpose is to use this PR to better understand the performance tests _as they run in the CI environment_. This may lead to a mergable PR but for now please ignore this if you come across it.

## What

The actual running of the performance tests accounts for almost all of the time that the GitHub Actions step takes for the performance test check. In this patch we're adding log points to report the individual runtimes for each step in that testing process, giving more insight into where our time is spent and where we might be able to address the poor runtime performance.

| Step | Time | % Overall |
|---|---|---|
| Fetch test branch | 50s (6s) | |
| Install deps | 4m 31s (3m 25s) | |
| Create env folders | 4ms (5ms) | |
| Prep PR branch | 5m 59s (4m 30s) | |
| Prep trunk branch | 5m 53s (4m 22s) | |
| Start env PR, post-editor | 1m 27s (1m 7s) | |
| Run test | 3m 28s (2m 46s) | |
| Stop env | 3s (3s) | |
| Start env trunk, post-editor | 49s (38s) | |
| Run test | 3m 42s (2m 37s) | |
| Stop env | 3s (2s) | |
| Start env PR, post-editor | 5s (4s) | |
| Run test | 3m 18s (2m 34s) | |
| Stop env | 3s (2s) | |
| Start env trunk, post-editor | 5s (4s) | |
| Run test | 3m 18s (2m 30s) | |
| Stop env | 3s (2s) | |
| Start env PR, post-editor | 5s (4s) | |
| Run test | 3m 20s (2m 36s) | |
| Stop env | 3s (2s) | |
| Start env trunk, post-editor | 5s (4s) | |
| Run test | 3m 26s (2m 29s) | |
| Stop env | 3s, 2s) | |
| Start env PR, site-editor | 5s (4s) | |
| Run test | 1m 27s (1m 7s) | |
| Stop env | 3s (2s) | |
| Start env trunk, site-editor | 5s (4s) | |
| Run test | 2m 3s (1m 7s) | |
| Stop env | 3s (2s) | |
| Start env PR, site-editor | 5s (4s) | |
| Run test | 1m 25s (59s) | |
| Stop env | 3s (2s) | |
| Start env trunk, site-editor | 5s (4s) | |
| Run test | 1m 42s (1m 14s) | |
| Stop env | 3s (2s) | |
| Start env PR, site-editor | 4s (4s) | |
| Run test | 1m 51s (1m 9s) | |
| Stop env | 3s (2s) | |
| Start env trunk, site-editor | 5s (4s) | |
| Run test | 1m 50s (1m 5s) | |
| Stop env | 3s (2s) | |

| Task | Total time | % total | % total _per_ |
|---|---|---|---|
| Install deps    | 4m 31s (3m 25s) | 8.3% | 8.3% |
| Fetch+Build branches | 12m (12m 42s) | 22% | 11% |
| Run post-editor | 21m (15m 32s) | 39% | 6.5% |
| Run site-editor | 10m (6m 41s) | 19% | 3.2% |

`post-editor` tests dominated by `firstContentfulPaint` and `firstBlock`, combined accounting for 80% of the test runtime.

## Why

Because the performance tests should not be the slowest thing holding us back from faster iteration speeds.